requirements.txt
# BotParty Robot Client
livekit>=1.0.0
livekit-api>=1.0.0
aiohttp>=3.11.0
pyyaml>=6.0.2
pydantic>=2.10.0
opencv-python-headless>=4.10.0
numpy>=2.0.0
Pillow>=11.0.0
pyserial>=3.5
paho-mqtt>=2.1.0
pyusb>=1.3.1
